requirements.txt
pyyaml>=6.0
structlog>=24.1.0
playwright>=1.48.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
pandas>=2.2.2
pyarrow>=17.0.0
numpy>=1.26.0
tqdm>=4.66.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
lightgbm>=4.0.0
openpyxl
langchain-google-genai
uvicorn
fastapi
nltk>=3.9
rank-bm25>=0.2.2
nltk>=3.9
